This commit adds functionality that will allow for a new strategy to be added to a player's set of strategies.
The add_strategy functionality is intended for use with games in strategic form. If it is applied to a game in extensive form it will raise a TypeError.

Additional changes have been made to the testing suite to verify the new functionality. Also, the player object documentation has been updated to reflect the new method.