This commit enhances the features of the action class.
Specifically, this commit adds the prob property to the action class. This can be used to read or set the property. It requires an int, decimal.Decimal, or fractions.Fraction item.

The property was added. Additionally, suitable tests were needed to ensure the funcitonality was working. Because of this a new test game was added that had action probabilities encoded.

Additionally, the python api documentation was updated. Also, this commit includes additional tests for node properties.